pyproject.toml
[project]
name = "mantra"
version = "0.1.0"
description = "Delaware Corporate Law AI Assistant - RAG-powered chatbot for legal research"
authors = [
    {name = "Mantra Team"}
]
requires-python = ">=3.9"
dependencies = [
    # Core LangChain
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    
    # Vector Database
    "faiss-cpu>=1.7.4",
    
    # OpenAI
    "openai>=1.12.0",
    
    # Web Framework
    "streamlit>=1.30.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.3,<2.0",
    
    # Document Processing
    "pypdf>=3.17.1",
    "unstructured>=0.10.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
